Create editedTextFile
This is the code I texted about. I wasn't sure how to get it in the format like the original code was. This was just copied and pasted from a text file. I just thought I would post it here if you guys want to view it or comment on it.